Separate page headings from actual content
We separate the page headings from the actual content, i.e. during
export we don't write them as content of the `<content>` elements, but
on import we merge them back in. This is important for any content
exchange with other CMSs which usually don't have the page headings as
part of the page contents, and it also simplifies the conversion
between different CMSimple_XH versions.

The drawback is that any additional markup of the headings will be
purged, but the usefulness of such markup is doubtful anyway, and if
the need arises we still could export the fully marked-up heading in an
additional `<page>` attribute.