Introduce Contents and Page entities
For now, these are mostly concerned about parsing and reconstructing
content.htm of CMSimple_XH ≥ 1.7.0 (new split mode).  We are going to
need these to

* be able to import old style content.htm (and maybe even very old
  style content.htm/pagedata.php, i.e. what we had prior to the
  "marriage"); we don't want to drop this feature, but still do not
  want support ancient CMSimple_XH versions directly
* to have meaningful and maintainable tests
* to possibly provide an API which allows to manipulate content.htm for
  some test environment setup, and maybe more[1]

The problem is that CMSimple_XH is in serious need of a full-fledged
contents model, something that unifies the bits and pieces (global
variables like `$c`, `$h`, `$u`, classes like `PageDataModel`, `Pages`,
`Publisher` and what not) into a single abstraction.  The new entities
are supposed to provide the basics of such a model.  While that may be
hard to maintain (unless integrated into the core), we feel it's worth
the effort.

[1] <https://cmsimpleforum.com/viewtopic.php?f=16&t=12663>