Tweak XML format
There is no need to export a `level` attribute, since this is implicit
by the page nesting, and we do no longer support irregular page
structures anymore (these are implicitly fixes when reading the
contents).

And we rename `pagedata` to `data`, because the former term is pretty
meaningless anyway.

Finally, we use a fixed version of `2.0`; using the CMSimple_XH version
does not really make sense; it's about the version of the XML schema.