Basic support for reading CMSimple_XH 1.6 content files
So far we only add `XhContents::fromXh16String()`, but don't actually
use it.  The code duplication with `::fromXhString()` is somewhat
annoying, but we can live with that for now.  More concerning is that
the `Document/DocumentStore` pattern breaks down (again); we already
used a couple of hacks to work-around the fact that no additional
parameters can be passed to the `::fromString()` methods, but regarding
the menu levels (which have been important for content splitting prior
to CMSimple_XH 1.7.0) the pattern breaks down.  We should reconsider
the use of repositories or something else.  Anyhow, for now we just
expose the `::fromXh16String()` method publicly, and expect it to be
called directly.