Proper internationalization for CSRF error message
This message is not shown to attacker, but rather to the Webmaster who
(almost) fell prey to the CSRF attack.  As such it makes sense to
provide proper internationalization (and also to mark up as error).
This is something CMSimple_XH didn't get right so far.

And yes, we are aware that same-site cookie restrictions can go a long
way to prevent CSRF attacks, but that would need to be supported by
CMSimple_XH (or manually configured by Webmasters), and there may still
be some loop holes, so we stick with our traditional prevention for the
time being.